weather_server/requirements.txt
requests==2.25.1
openmeteo_requests
requests_cache
retry_requests
psycopg2-binary
numpy
scikeras
keras==3.0.0
scikit-learn
tensorflow>=2.7.0
pandas
datetime